Remove FilterBlobSerializer dependents from plugin
FilterSerializer and validate_jsonfilter depend on FilterBlobSerializer
and need not themselves be pluginable.
