+++ src/argus/notificationprofile/serializers.py
@@ -1,14 +1,10 @@
 from rest_framework import fields, serializers
 
-from argus.filter import get_filter_backend
-from argus.filter.primitive_serializers import CustomMultipleChoiceField
+from argus.filter.serializers import FilterSerializer
 from argus.incident.constants import INCIDENT_LEVELS
 from argus.incident.models import Event
 from .media import api_safely_get_medium_object
 from .models import DestinationConfig, Media, NotificationProfile, TimeRecurrence, Timeslot
-
-filter_backend = get_filter_backend()
-FilterSerializer = filter_backend.FilterSerializer
 
 
 class TimeRecurrenceSerializer(serializers.ModelSerializer):
