+++ src/argus/notificationprofile/views.py
@@ -15,6 +15,7 @@ from drf_rw_serializers import viewsets as rw_viewsets
 
 from argus.drf.permissions import IsOwner
 from argus.filter import get_filter_backend
+from argus.filter.serializers import FilterSerializer
 from argus.incident.serializers import IncidentSerializer
 from argus.notificationprofile.media import api_safely_get_medium_object
 from argus.notificationprofile.media.base import NotificationMedium
@@ -33,7 +34,6 @@ from .serializers import (
 
 filter_backend = get_filter_backend()
 QuerySetFilter = filter_backend.QuerySetFilter
-FilterSerializer = filter_backend.FilterSerializer
 FilterBlobSerializer = filter_backend.FilterBlobSerializer
